+++ src/lib/terrain/voxelmap/viewer/voxelmap-viewer.ts
@@ -4,9 +4,9 @@ import { vec3ToString } from '../../../helpers/string';
 import * as THREE from '../../../libs/three-usage';
 import { type MaterialsStore } from '../../materials-store';
 import { type VoxelsChunkOrdering, type VoxelsChunkSize } from '../i-voxelmap';
-import { ChunkRenderableFactoryCpu } from '../chunk/chunkRenderableFactory/chunk-renderable-factory-cpu';
-import { ChunkRenderableFactoryCpuWorker } from '../chunk/chunkRenderableFactory/chunk-renderable-factory-cpu-worker';
-import { ChunkRenderableFactoryGpu } from '../chunk/chunkRenderableFactory/chunk-renderable-facory-gpu';
+import { ChunkRenderableFactoryCpu } from '../chunk/chunkRenderableFactory/cpu/chunk-renderable-factory-cpu';
+import { ChunkRenderableFactoryCpuWorker } from '../chunk/chunkRenderableFactory/cpu/chunk-renderable-factory-cpu-worker';
+import { ChunkRenderableFactoryGpu } from '../chunk/chunkRenderableFactory/gpu/chunk-renderable-facory-gpu';
 import { type ChunkRenderableFactoryBase } from '../chunk/chunkRenderableFactory/chunk-renderable-factory-base';
 import { ChunkId } from '../chunk/chunk-id';
 import { type CheckerboardType, type VoxelsChunkData } from '../voxelsRenderable/voxelsRenderableFactory/voxels-renderable-factory-base';
